Make parameters optional (except for output-directory)
Some of the parameters are now optional in case the devs want to use the config in their .slather.yml file. The output-directory param will have to be fixed since the script for reading the coverage file will have to be known in Jenkins -- where there's another script that actually performs the Sonar operation).